Trait glifparser::component::FlattenedGlif

pub trait FlattenedGlifwhere
    Self: Clone,{
    fn flattened(
        &self,
        rects: &mut Option<Vec<ComponentRect>>
    ) -> Result<Self, GlifParserError>;
}

Required Methods

fn flattened(
&self,
rects: &mut Option<Vec<ComponentRect>>
) -> Result<Self, GlifParserError>

Check that all components in your .glif file really resolve, and if they do, get their contours and apply their matrices. If you want bounding rectangles as this process is done with a logical name for each rectangle you can draw, pass in rects as a &mut