Module glifparser::glif

Re-exports

pub use self::xml::XMLConversion;

Modules

xml

Structs

Glif

A UFO .glif

Enums

Lib

Traits

FromXML

GlifLike

IntoXML

Functions

name_to_filename

read

Read UFO .glif XML to Glif struct. This should only be used when you have no known filename, and the glif is unattached from a UFO.

read_from_filename

If you have a known filename, it is always preferable to call this function, as it sets the filename on the Glif as well as on its GlifComponent’s, easing their transition into Component’s.

read_from_filename_pedantic

read_pedantic

write

write_to_filename