Module glifparser::outline::contour

Enums

End

Traits

CheckSmooth

This validates the UFO .glif point smooth attribute.

GenericPrevNext

Generic prev/next given an index for Vec<T>.

PrevNext

Return the previous and next index, given an index. As opposed to GenericPrevNext, this always considers a contour’s open/closed state (assert!(self[0].ptype == PointType::Move)).

Reverse

Reverse the logical (visual) order of Bézier splines in a contour, flipping handles as necessary, and taking into account open/closed state.

State

Is contour open or closed?