Module glifparser::outline::skia

Structs

SkiaPaths

glifparser returns for you two Skia paths when called on an outline, because it is extremely liekly that you are going to want to draw open paths in a different way than you draw closed ones. Option’s are used because it’s possible, and most likely, that the Outline will only contain paths of one type.

SkiaPointTransforms

Traits

ConicsToCubics

FromSkOutline

FromSkiaPath

Get an outline from a Skia path. Outline is guaranteed to contain only Curve’s, no QCurve’s.

SplitSkiaPath

ToSkiaPath

ToSkiaPaths

TryIntoPointType

Type Definitions

SkContour

SkOutline

SkPointTuple