Trait glifparser::outline::RoundToInt

pub trait RoundToInt {
    fn round_to_int(&mut self);
}

This trait is primarily intended for easing .glif equality testing internally by our test suite. It therefore doesn’t do any of the fancy things it could like change point types and assert handles as colocated. Consider MFEKmath::Fixup, RefigurePointTypes, etc., not this. (…Or perhaps together?)

Required Methods

fn round_to_int(&mut self)