Module glifparser::point

.glif <point>

Structs

GlifPoint

A “close to the source” .glif <point>

KurboPoint

A 2D point.

Point

A Skia-friendly point

Enums

Handle

A handle on a point

PointType

WhichHandle

For use by Point::handle_or_colocated TODO: Replace with Option

Traits

FromKurboPoint

IsValid

PointData

API consumers may put any clonable type as an associated type to Glif, which will appear along with each Point. You could use this to implement, e.g., hyperbeziers. The Glif Point’s would still represent a Bézier curve, but you could put hyperbezier info along with the Point.

PointLike

ToKurboPoint