Trait glifparser::outline::contour::GenericPrevNext

pub trait GenericPrevNext {
    type Error;
    fn prev_next(&self, idx: usize) -> Result<(usize, usize), Self::Error>;

    fn idx_which_end(&self, idx: usize) -> Result<Option<End>, Self::Error>;

    fn idx_sane(&self, idx: usize) -> Result<(), Self::Error>;

    fn prev(&self, idx: usize) -> Result<usize, Self::Error> { ... }

    fn next(&self, idx: usize) -> Result<usize, Self::Error> { ... }

    fn idx_at_start_or_end(&self, idx: usize) -> Result<bool, Self::Error> { ... }

    fn idx_is_sane(&self, idx: usize) -> bool { ... }

    fn idx_is_insane(&self, idx: usize) -> bool { ... }
}

Associated Types

type Error

Required methods

fn prev_next(&self, idx: usize) -> Result<(usize, usize), Self::Error>

Return the previous and next index, given an index.

fn idx_which_end(&self, idx: usize) -> Result<Option<End>, Self::Error>

Returns which end we have, given an index.

fn idx_sane(&self, idx: usize) -> Result<(), Self::Error>

Returns whether the index is sane given the vec it’s called on, intended to be used with ? shortcut.

Provided methods

fn prev(&self, idx: usize) -> Result<usize, Self::Error>

fn next(&self, idx: usize) -> Result<usize, Self::Error>

fn idx_at_start_or_end(&self, idx: usize) -> Result<bool, Self::Error>

fn idx_is_sane(&self, idx: usize) -> bool

fn idx_is_insane(&self, idx: usize) -> bool