Struct glifparser::point::Point

pub struct Point<PD: PointData> {
    pub x: f32,
    pub y: f32,
    pub a: Handle,
    pub b: Handle,
    pub name: Option<String>,
    pub ptype: PointType,
    pub smooth: bool,
    pub data: Option<PD>,
}

A Skia-friendly point

Fields
x: f32y: f32a: Handleb: Handlename: Option<String>ptype: PointTypesmooth: booldata: Option<PD>
Implementations

impl<PD: PointData> Point<PD>

pub fn apply_matrix(&mut self, matrix: Affine)

impl<PD: PointData> Point<PD>

pub fn new() -> Point<PD>

pub fn from_x_y_type((x, y): (f32, f32), ptype: PointType) -> Point<PD>

Make a point from its x and y position and type

pub fn from_x_y_a_b_type(
(x, y): (f32, f32),
(a, b): (Handle, Handle),
ptype: PointType
) -> Point<PD>

Make a point from its x and y position, handles and type

pub fn from_fields(
(x, y): (f32, f32),
(a, b): (Handle, Handle),
smooth: bool,
ptype: PointType,
name: Option<String>,
data: Option<PD>
) -> Point<PD>

Make a point from its x and y position, handles and type

pub fn handle(&self, which: WhichHandle) -> Handle

pub fn handle_or_colocated(
&self,
which: WhichHandle,
transform_x: &dyn Fn(f32) -> f32,
transform_y: &dyn Fn(f32) -> f32
) -> (f32, f32)

Return an x, y position for a point, or one of its handles. If called with WhichHandle::Neither, return position for point.

pub fn handle_as_gpoint(&self, which: WhichHandle) -> GlifPoint

pub fn handle_as_point(&self, which: WhichHandle) -> Self

pub fn handle_as_kpoint(&self, which: WhichHandle) -> KurboPoint

pub fn set_handle(&mut self, which: WhichHandle, handle: Handle)

This function is intended for use by generic functions that can work on either handle, to decrease the use of macros like move_mirror!(a, b).